Start creating initializer interface.
